Geo_layout_codebase/Geo_LayoutLM-geo_code_latest_timeoptm_version_updated_apr22_sharing_backend/training_main/main/requirements_final_idp.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albumentations==1.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-core==1.26.4
azure-storage-blob==12.16.0
beautifulsoup4==4.12.2
boto3==1.34.83
botocore==1.34.162
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
datasets==2.17.1
dill==0.3.8
dnspython==2.7.0
dropbox==11.36.2
elasticsearch==6.8.2
et_xmlfile==2.0.0
fastapi==0.110.2
filelock==3.17.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2023.10.0
gensim==4.3.1
google-api-core==2.24.1
google-api-python-client==2.86.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.4.3
googleapis-common-protos==1.59.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.28.1
idna==3.10
imagededup==0.3.1
imageio==2.37.0
isodate==0.7.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
langid==1.1.6
lazy_loader==0.4
lightning-utilities==0.11.9
lit==18.1.8
MarkupSafe==3.0.2
matplotlib==3.10.0
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
overrides==3.1.0
packaging==24.2
pandas==2.2.3
pillow==10.3.0
piq==0.7.1
ply==3.11
pretrainedmodels==0.7.4
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psutil==5.9.7
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.6.3
PyMuPDF==1.23.9
PyMuPDFb==1.23.9
pynvml==11.5.0
pyparsing==3.2.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.2.1
pytz==2023.3
PyWavelets==1.8.0
PyYAML==6.0
qudida==0.0.4
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.2
scikit-image==0.21.0
scikit-learn==1.6.1
scipy==1.10.1
seaborn==0.13.2
seqeval==1.2.2
six==1.17.0
sklearn-pandas==2.2.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.37.2
stone==3.3.8
sympy==1.13.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==0.9.12
tokenizers==0.21.0
torch==2.0.1
torchmetrics==1.6.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.48.2
triton==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.1.15
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.22.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3